+++ src/users/users.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { User } from './user.entity';
+import { User } from './users.entity';
 
 @Module({
     imports: [TypeOrmModule.forFeature([User])]
 })
-export class UserModule {}
+export class UsersModule {}
